% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_fitting.R
\name{cellwise_influences}
\alias{cellwise_influences}
\title{Evaluates where ligands are most strongly driving trajectory dynamics.}
\usage{
cellwise_influences(
  obj,
  ligands_list = NULL,
  ligand_target_matrix,
  slot = "data",
  window_pct = 0.1,
  step_size = 0.02,
  n_top_genes = 100,
  n_top_ligands = 5
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with get_path_ligands or get_traj_ligands_monocle.}

\item{slot}{Slot of Seurat object from which to retrieve expression data.}

\item{window_pct}{Size of the window, as a percentage of the number of cells in a path.}

\item{step_size}{Size of the step, or interval, as a percentage of the number of cells in a path.}

\item{n_top_genes}{Number of top importance genes to consider in the cellwise analysis. Because we have already evaluated the ligand importances for a branch, we restrict our more granular analysis to the top genes and ligands.}

\item{n_top_ligands}{Number of top ligands to consider in the cellwise analysis.}
}
\value{
'a Seurat object updated with per-cell ligand influence score in obj@misc$entrain$path$path_name$cellwise_influences.
}
\description{
Performs a sliding window analysis along the trajectory to identify windows where environmental influence is most apparent.
}
