% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{draw_entrain_circos}
\alias{draw_entrain_circos}
\title{Draw circos plot of Entrain ligands.}
\usage{
draw_entrain_circos(
  obj,
  path_name = NULL,
  vcluster = NULL,
  n_top_ligands = 10,
  lr_network,
  weighted_networks,
  palette = "Spectral"
)
}
\arguments{
\item{obj}{A seurat object containing the Entrain result from entrain_velocity or get_traj_ligands()}

\item{path_name}{Optional. String indicating the name of the trajectory path. Use this if you want to plot ligands associated with pseudotime branches.}

\item{vcluster}{Optional. String indicating the name of the velocity cluster. Use this if you wish to plot ligands associated with velocity clusters.}

\item{n_top_ligands}{Number of top ligands to plot}

\item{lr_network}{NicheNet ligand-receptor pairs data file.}

\item{weighted_networks}{NicheNet weighted networks.}
}
\value{
NULL. Plots the circlize object to graphics output.
}
\description{
'Given a seurat object containing Entrain results, draw the ligands driving the observed velocities or trajectories..
}
