% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_fitting.R
\name{get_path_ligands}
\alias{get_path_ligands}
\title{Core function for Entrain ligand assignment that is agnostic to the trajectory inference algorithm used.}
\usage{
get_path_ligands(
  obj,
  expressed_ligands,
  lr_network,
  ligand_target_matrix,
  path_cell_names,
  pseudotime_key = "pseudotime",
  cluster_key = NULL,
  reduction_name = "MonocleUMAP",
  covariance_cutoff = 0.05,
  return_model = FALSE
)
}
\arguments{
\item{obj}{A seurat object  with cell pseudotimes contained in a metadata column with column name pseudotime_key.}

\item{expressed_ligands}{Character vector of active ligands in dataset (must be in rownames(obj))}

\item{lr_network}{NicheNet ligand-receptor pairs data file.}

\item{ligand_target_matrix}{NicheNet ligand-target data file.}

\item{path_cell_names}{a vector of cell names (must be in colnames(obj)), defining cells in a trajectory path for which associated ligands will be calculated.}

\item{pseudotime_key}{Column name in colnames(obj@meta.data) that contains cell pseudotime values.}

\item{cluster_key}{Column name in colnames(obj@meta.data) corresponding to}

\item{reduction_name}{Seurat reduction key for dimension reduction visualization.}

\item{covariance_cutoff}{Remove the bottom \code{covariance_cutoff} fraction of covariances (e.g. 0.10 = bottom 10 percent of genes, ranked by covariance, removed from later analysis).}

\item{return_model}{If TRUE, will additionally save the raw random forest object in obj@misc$entrain$paths$path_name$model}
}
\value{
a Seurat object with ligand trajectory results in obj$misc$entrain$paths$path_name
}
\description{
Runs Entrain to identify ligands responsible for trajectory dynamics. Requires pseudotime labels for each cell in a column in metadata with name pseudotime_key
}
