% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_fitting.R
\name{get_senders_interactive}
\alias{get_senders_interactive}
\title{Interactive selection of sender clusters.}
\usage{
get_senders_interactive(
  sender_obj,
  sender_cluster_key,
  reduction_key = NULL,
  lr_network
)
}
\arguments{
\item{sender_obj}{Seurat object containing niche cell data}

\item{sender_cluster_key}{column name in sender_obj@meta.data that denotes the clusters.}

\item{reduction_key}{Seurat reduction key for dimension reduction visualization.}

\item{lr_network}{NicheNet ligand-receptor pairs data file.}
}
\value{
A vector of cluster names.
}
\description{
Choosing potential sender cell clusters in a GUI. Inspired from Monocle3's interactive workflow.
}
