% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_fitting.R
\name{get_traj_ligands_monocle}
\alias{get_traj_ligands_monocle}
\title{Identify ligands responsible for trajectory dynamics using Monocle graph learning.}
\usage{
get_traj_ligands_monocle(
  obj,
  sender_obj = NULL,
  cds = NULL,
  sender_cluster_key = NULL,
  sender_cluster_names = NULL,
  num_dim = 10,
  root_cells = NULL,
  root_pr_nodes = NULL,
  expressed_ligands = NULL,
  path_cell_names = NULL,
  path_nodes = NULL,
  minimal_branch_len = 5,
  ncenter = 50,
  use_partition = TRUE,
  precomputed_umap = NULL,
  prune_graph = T,
  close_loop = F,
  overwrite = FALSE,
  reduction_key = "MonocleUMAP_",
  sender_reduction_key = "umap",
  expression_proportion_cutoff = 0.01,
  covariance_cutoff = 0.05,
  lr_network,
  ligand_target_matrix,
  export_cds = TRUE
)
}
\arguments{
\item{obj}{A seurat object containing receiver cells undergoing a trajectory}

\item{sender_obj}{A seurat obj containing sender cells expressing ligands. Set to NULL if your sender cells are contained in 'obj'.}

\item{cds}{Optional. You can supply a \code{monocle3} \code{cds} object here if you have run \code{monocle3::order_cells()} on it already.}

\item{sender_cluster_key}{Column name in metadata that denotes where sender cells are specified. If NULL, this defaults to the default Seurat Idents given by \code{levels(obj)}}

\item{sender_cluster_names}{Unique values in column \code{sender_cluster_key} that denote sender cell clusters.}

\item{num_dim}{Number of dimensions for monocle reduce_dimension function.}

\item{root_cells}{Optional. Character vector of root cell names that comprise the beginning of the trajectory (must be in colnames(obj)). If not given or NULL, will open a window to interactively select a root cell (Monocle's default behaviour).}

\item{root_pr_nodes}{Argument for \code{monocle3::order_cells()}. Use if you want to specify root nodes programmatically.}

\item{expressed_ligands}{Character vector of active ligands in dataset (must be in rownames(obj))}

\item{path_cell_names}{Optional. Character vector of cell names that comprise the differentiation trajectory to be analyzed by Entrain (must be in colnames(obj)). If not given or NULL, will open a window to interactively select the trajectory path to analyze (Monocle's default behaviour).}

\item{path_nodes}{Use if you want to specify the path without user interaction. e.g. \code{path_nodes = c("Y_2", "Y_50")} tells Entrain to analyze cells between the nodes \code{Y_2} and \code{Y_50}}

\item{minimal_branch_len}{Parameter for Monocle graph learning.}

\item{ncenter}{Parameter for Monocle graph learning.}

\item{use_partition}{Parameter for Monocle graph learning}

\item{precomputed_umap}{If you have a precomputed UMAP on your receiver cells that you wish to build a monocle trajectory on, then set this to the name of your precomputed embedding e.g. precomputed_umap = 'umap'}

\item{prune_graph}{Parameter for Monocle graph learning.}

\item{close_loop}{Parameter for Monocle graph learning.}

\item{overwrite}{If there is an existing Monocle trajectory in the obj (e.g. if you have run this function before), overwrite = TRUE will delete the old trajectory. Otherwise, the old trajectory will be used for Entrain.}

\item{reduction_key}{Seurat reduction key for dimension reduction visualization.}

\item{sender_reduction_key}{Only applicable if sender_obj is not NULL: The dimension reduction key of sender_obj that you wish to use,}

\item{expression_proportion_cutoff}{Pct cutoff to threshold a ligand as active. A ligand is 'active' if it is expressed in more than \code{expression_proportion_cutoff} fraction of cells in the sender cluster.}

\item{covariance_cutoff}{Remove the bottom \code{covariance_cutoff} fraction of covariances (e.g. 0.10 = bottom 10 percent of genes, ranked by covariance, removed from later analysis).}

\item{lr_network}{NicheNet ligand-receptor pairs data file.}

\item{ligand_target_matrix}{NicheNet ligand-target data file.}

\item{export_cds}{If TRUE, saves the Monocle cell_data_set trajectory object in obj@misc$monocle_graph. Set TRUE if you want to plot Entrain results afterwards.}
}
\value{
a Seurat object with ligand velocity results in obj$misc$entrain$velocity_result.
}
\description{
Runs a standard Monocle3 trajectory reconstruction pipeline, followed by Entrain to identify ligands responsible for trajectory dynamics.
}
