% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_velocity.R
\name{get_velocity_ligands}
\alias{get_velocity_ligands}
\title{'Identify ligands responsible for observed velocity vectors in velocity clusters.}
\usage{
get_velocity_ligands(
  obj,
  fit_likelihoods,
  velocity_clusters,
  expressed_ligands,
  expression_proportion_cutoff,
  lr_network,
  ligand_target_matrix
)
}
\arguments{
\item{obj}{A seurat object containing the receiver cells undergoing a differentiation trajectory}

\item{lr_network}{NicheNet ligand-receptor pairs data file.}

\item{ligand_target_matrix}{NicheNet ligand-target data file.}

\item{adata_clustered}{An anndata file containing velocities, preclustered with cluster_velocities().}

\item{save_adata}{Optional. Required if want to plot your output. Filename of anndata object to write results to. If not given or NULL, anndata will not be saved. Recommended if you want to re-run or continue analysis in scanpy/scvelo.}

\item{cluster_key}{Column name of the cluster labels in the metadata of adata_clustered.}
}
\value{
'a Seurat object with ligand velocity results in obj$misc$entrain$velocity_result.
}
\description{
'Given an anndata object containing clustered velocity vectors, and the dynamical model for each gene obtained with scvelo recover_dynamics
identify ligands responsible for driving the top likelihood dynamics genes.
}
