% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_fitting.R
\name{intersect_genes}
\alias{intersect_genes}
\title{Intersect seurat object genes}
\usage{
intersect_genes(obj, h5ad)
}
\arguments{
\item{obj}{Seurat object to be subsetted}

\item{h5ad}{String denoting filename of h5ad from with to intersect genes with.}
}
\value{
A seurat object with features subsetted to the intersection of genes in obj and h5ad.
}
\description{
Subsets the genes of obj to the genes that are the intersection of obj and the anndata h5ad
}
