% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_covarying_genes_heatmap}
\alias{plot_covarying_genes_heatmap}
\title{Heatmap plot of top trajectory gene covariances.}
\usage{
plot_covarying_genes_heatmap(obj, path = paths[1], n_top_genes = 10)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with get_path_ligands or get_traj_ligands_monocle.}

\item{path}{Name of list element in obj@misc$entrain$paths corresponding to the trajectory path that you wish to visualize.}

\item{n_top_genes}{Number of top covarying genes to visualize. Default 10.}
}
\value{
A ggplot2 plot object.
}
\description{
Recovers the covariance of the top genes that covary with pseudotime and plots it as a 1d-heatmap.
}
