% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_covarying_genes_scatter}
\alias{plot_covarying_genes_scatter}
\title{' Scatter plot of top trajectory covarying genes.}
\usage{
plot_covarying_genes_scatter(
  obj,
  path = paths[1],
  n_top_genes = 10,
  color_key = NULL,
  nrow = 5,
  ncol = 2
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with get_path_ligands or get_traj_ligands_monocle.}

\item{path}{Name of list element in obj@misc$entrain$paths corresponding to the trajectory path that you wish to visualize.}

\item{n_top_genes}{Number of top importance target genes to visualize. Default 10.}

\item{color_key}{A column name in obj@meta.data corresponding to a label you wish to color by e.g., a cluster label.}

\item{nrow}{Number of rows for the facet plot grid.}

\item{ncol}{Number of columns for the facet plot grid.}
}
\value{
'A ggplot2 plot object.
}
\description{
Recovers the expression of the top genes that covary with pseudotime and plots it as a scatter plot.
}
