% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_ligand_influences}
\alias{plot_ligand_influences}
\title{Plots ligand cellwise inlfuences on a UMAP.}
\usage{
plot_ligand_influences(
  obj,
  colormap = "Set1",
  reduction = "MonocleUMAP_",
  custom_colors = NULL,
  ligand_labels = TRUE,
  n_top_ligands = 5,
  label_font_size = 4,
  label_position = "endpoint"
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with get_path_ligands or get_traj_ligands_monocle.}

\item{colormap}{String denoting a brewer.pal compatible color palette.}

\item{reduction}{Name of reduction to be used for plotting. Defaults to "MonocleUMAP"; the UMAP coordinates generated by the Monocle workflow.}

\item{custom_colors}{Optional vector of colors to be passed to ggplot. Will override colormap parameter.}
}
\value{
A ggplot object with cells plotted on a UMAP with cell color corresponding to the degree of ligand influence for a particular branch.
}
\description{
Requires cellwise_influences to be run on obj beforehand.
}
