% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_ligand_targets}
\alias{plot_ligand_targets}
\title{Plots a heatmap of ligand-target regulatory potentials extracted from NicheNet.}
\usage{
plot_ligand_targets(
  obj,
  path,
  n_top_targets = 10,
  n_top_ligands = 10,
  ligand_target_matrix
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with get_path_ligands or get_traj_ligands_monocle.}

\item{n_top_targets}{Number of top importance target genes to visualize. Default 10.}

\item{n_top_ligands}{Number of top ligands to visualize. Default 10.}

\item{ligand_target_matrix}{NicheNet ligand-target data file.}
}
\value{
A plotly plot object.
}
\description{
Visualize the ligand-target regulatory relationships that are driving the analyzed trajectory paths.
}
