% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_sender_influence_velocity}
\alias{plot_sender_influence_velocity}
\title{Plot sender influences overlayed on a RNA velocity plot}
\usage{
plot_sender_influence_velocity(
  adata,
  velocity_clusters,
  velocity_cluster_key = "vcluster",
  frameon = FALSE,
  size = 40,
  sender_palette = "Reds",
  title = " ",
  filename = "entrain_velocity_influence.png"
)
}
\arguments{
\item{adata}{Path to a \code{.h5ad} \code{anndata} object. Must have had entrain_velocity run on it.}

\item{velocity_clusters}{Values in \code{adata.obs.velocity_cluster_key} denoting which velocity clusters to visualize. The sender influences will represent contribution towards this velocity cluster only, not all velocities in the dataset.}

\item{velocity_cluster_key}{Column name of \code{adata.obs} metadata denoting velocity clusters.}

\item{frameon}{scvelo.pl.velocity_embedding argument: Draws a frame around the plot.}

\item{size}{scvelo.pl.velocity_embedding argument: Size of points.}

\item{sender_palette}{A matplotlib colorscale.}

\item{title}{Plot title.}

\item{filename}{Filename of plot to save.}
}
\value{
NULL. Saves a plot in working directory with name \code{filename}
}
\description{
Visualizes the contribution of each cell in the dataset towards the observed RNA velocities.
}
