% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_velocity_likelihoods_heatmap}
\alias{plot_velocity_likelihoods_heatmap}
\title{Plot velocity likelihoods as a heatmap.}
\usage{
plot_velocity_likelihoods_heatmap(
  obj,
  velocity_clusters = NULL,
  n_top_genes = 5,
  color_low = "#f6ffff",
  color_high = "#01bfc4",
  colorscale = "BuPu"
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with entrain_velocity}

\item{velocity_clusters}{Character vector of velocity clusters to include in the plot. Default \code{NULL} - include all velocity clusters,}

\item{n_top_genes}{Number of top likelihood  genes to visualize. Default 10.}

\item{color_low}{Hex code of lower limit}

\item{color_high}{Hex code of upper limit.}

\item{colorscale}{Name of a color palette to be passed to the palette argument of scale_fill_distiller. Overrides color_low and color_high.}
}
\value{
A gglot2 plot object.
}
\description{
Visualize the top gene likelihoods for each velocity clusters.
}
