% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_velocity.R
\name{entrain_velocity}
\alias{entrain_velocity}
\title{Identify ligands responsible for observed velocity vectors.}
\usage{
entrain_velocity(
  receiver_obj,
  sender_obj = NULL,
  adata,
  expressed_ligands = NULL,
  sender_cluster_key = NULL,
  sender_clusters = NULL,
  reduction_key = NULL,
  save_adata = NULL,
  velocity_cluster_key = NULL,
  num_jobs = 10L,
  num_top_genes = 500L,
  expression_proportion_cutoff = 0.1,
  lr_network = NULL,
  ligand_target_matrix = NULL,
  ...
)
}
\arguments{
\item{save_adata}{Optional. Filename of anndata object to write results to. If not given or NULL, anndata will not be saved. Recommended if you want to re-run or continue analysis in scanpy/scvelo.}

\item{velocity_cluster_key}{Column name of the cluster labels in the metadata of adata_clustered. Can be calculated manually in scvelo with scvelo.tl.velocity_clusters().}

\item{lr_network}{NicheNet ligand-receptor pairs data file.}

\item{ligand_target_matrix}{NicheNet ligand-target data file.}

\item{obj}{A seurat object}

\item{adata_clustered}{An anndata file containing velocities, preclustered with cluster_velocities().}

\item{n_jobs}{Long integer. Number of parallel jobs to run for scvelo recover dynamics function.}

\item{n_top_genes}{Long integer. Number of top velocity genes to calculate likelihoods for. Default 500L.}

\item{resolution}{Optional argument defining resolution of velocity clustering. Default 0.05.}
}
\value{
'a Seurat object with ligand velocity results in obj$misc$entrain$velocity_result.
}
\description{
Given an anndata object containing clustered velocity vectors, and NicheNet ligand database, identify ligands responsible for velocity vectors.
This function first calls a wrapper for scvelo's recover_dynamics function, obtaining gene-wise velocity dynamics for each cluster. Folloewd by Entrain ligand analysis.
Note that velocity clusters are generated from velocity data (i.e. time derivative of expression), rather than expression data like a typical usual cell clustering workflow such as Seurat FindClusters.
Those wishing to run this on their own pre-generated clusters (E.g. if you have a bespoke set of marker genes) can do so by inputting an anndata with the custom cluster annotations
in the anndata.obs metadata, and passing the column name to cluster_key But this is not quite recommended because cells that are similar in expression could exhibit very different velocities. In turn,
this will input a noisy dataset to Entrain's randomForest, and any ligands that come out of such an analysis will be spurious.
}
