% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_ligand_trajectories}
\alias{plot_ligand_trajectories}
\title{Plots ligand influences and variance explained scores on a Monocle trajectory plot.}
\usage{
plot_ligand_trajectories(
  obj,
  color_cells_by = "pseudotime",
  n_top_ligands = 5,
  trajectory_graph_color = "grey70",
  group_label_size = 5,
  trajectory_graph_segment_size = 1,
  cell_size = 0.6,
  label_groups_by_cluster = FALSE,
  label_font_size = 4,
  cds = NULL
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with get_path_ligands or get_traj_ligands_monocle.}

\item{color_cells_by}{Argument for \code{monocle3::plot_cells()}. A column in \code{colData(obj@misc$entrain$monocle_cds)} or \code{colData(cds)} denoting how to color the cells. Default \code{pseudotime}}

\item{n_top_ligands}{Number of top ligands to write on each plot label.}

\item{trajectory_graph_color}{Parameters to be passed to Monocle plot_cells function.}

\item{group_label_size}{Parameters to be passed to Monocle plot_cells function.}

\item{trajectory_graph_segment_size}{Parameters to be passed to Monocle plot_cells function.}

\item{cell_size}{Parameters to be passed to Monocle plot_cells function.}

\item{label_groups_by_cluster}{Argument for \code{monocle3::plot_cells()}. Default FALSE.}

\item{label_font_size}{Argument for \code{monocle3::plot_cells()}. Default 4.}

\item{cds}{Optional cell_data_set object to use instead of the default cds at \code{obj@misc$entrain$monocle_cds}}
}
\value{
a ggplot object with cells, trajectories, and ligand labels plotted on a UMAP.
}
\description{
Requires get_path_ligands or get_traj_ligands_monocle to be run on dataset beforehand. Specifically for plotting Entrain results analyzed on a Monocle trajectory.
}
