% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_velocity_ligand_targets}
\alias{plot_velocity_ligand_targets}
\title{Plots a heatmap of ligand-target regulatory potentials extracted from NicheNet.}
\usage{
plot_velocity_ligand_targets(
  obj,
  n_top_targets = 10,
  velocity_cluster = NULL,
  n_top_ligands = 10,
  ligand_target_matrix,
  color_low = "#f6ffff",
  color_high = "#01bfc4",
  colorscale = "Greens"
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with entrain_velocity}

\item{n_top_targets}{Number of top likelihood target genes to visualize. Default 10.}

\item{velocity_cluster}{Name of the velocity cluster whose regulatory relationships you wish to plot.}

\item{n_top_ligands}{Number of top ligands to visualize. Default 10.}

\item{ligand_target_matrix}{NicheNet ligand-target data file.}

\item{color_low}{Hex code of lower limit}

\item{color_high}{Hex code of upper limit.}

\item{colorscale}{Name of an RColorBrewer palette. Default \code{Greens}}
}
\value{
A gglot2 plot object.
}
\description{
'Visualize the ligand-target regulatory relationships that are driving the analyzed velocities.
}
