% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{plot_velocity_ligands}
\alias{plot_velocity_ligands}
\title{Plot top ligands overlayed on a RNA velocity plot}
\usage{
plot_velocity_ligands(
  adata,
  color = "dimgrey",
  alpha = 0.1,
  density = 3,
  arrow_size = 1,
  linewidth_scale = 2,
  vector_type = "stream",
  velocity_cluster_key = "vcluster",
  velocity_clusters = NULL,
  cell_palette = NULL,
  size = 100,
  velocity_cluster_palette = "gnuplot",
  label_fontsize = 15,
  top_n_ligands = 5L,
  figsize = c(10, 10),
  plot_output_path = "entrain_velocity_ligands.png"
)
}
\arguments{
\item{adata}{Path to a \code{.h5ad} \code{anndata} object. Must have had entrain_velocity run on it.}

\item{color}{scvelo.pl argument: Key for annotations of observations/cells or variables/genes. Use if you want dot color to denote a cell type or other annotation. Default is \code{NULL} - all cells are colored a single color denoted by argument \code{cell_palette}}

\item{alpha}{Opacity of points.}

\item{density}{Density of RNA velocity vectors.}

\item{arrow_size}{Size of RNA velocity vector arrowhead.}

\item{linewidth_scale}{Width of RNA velocity vector body.}

\item{vector_type}{Whether to plot a \code{scvelo.pl.velocity_embedding_stream()} plot or a \code{scvelo.pl.velocity_embedding_grid()} plot. Default \code{stream}}

\item{velocity_cluster_key}{Column name of \code{adata.obs} metadata denoting velocity clusters.}

\item{velocity_clusters}{Values in \code{adata.obs.velocity_cluster_key} denoting which velocity clusters and their ligands to visualize. Defaults to NULL, in which it will plot the ligands for all positive variance explained velocity clusters}

\item{cell_palette}{A matplotlib color palette for annotations denoted in argument \code{color}.}

\item{size}{scvelo.pl.velocity_embedding argument: Size of points.}

\item{velocity_cluster_palette}{A matplotlib color palette denoting colorscale of velocity clusters e.g., \code{"Spectral"}. OR a single color in \code{matplotlib.colors.CSS4_COLORS.keys()} e.g. \code{"black"};  OR: A vector of hex codes with the same length as the number of velocity clusters being plotted e.g., \code{c("#bfe5a0", "#9e0242", "#d8434e", "#f67b4a", "#5e4fa2")}.}

\item{label_fontsize}{Font size of ligand labels}

\item{top_n_ligands}{Number of top ligands to include in each label.}

\item{figsize}{Figure size. Default is \code{c(10,10)}}

\item{plot_output_path}{Filename of plot to save.}
}
\value{
NULL. Saves a plot in working directory with name \code{plot_output_path}
}
\description{
Visualizes the ligands most likely to be driving the observed RNA velocities.
}
