% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrain_plotting.R
\name{velocity_ligand_importance_heatmap}
\alias{velocity_ligand_importance_heatmap}
\title{velocity_ligand_importance_heatmap}
\usage{
velocity_ligand_importance_heatmap(
  obj,
  vclusters = NULL,
  n_top_ligands = 5,
  ligand_target_matrix,
  color_low = "#f6ffff",
  color_high = "#01bfc4",
  colorscale = "BuPu",
  rescale_columns = TRUE
)
}
\arguments{
\item{obj}{A seurat object that has been analyzed with entrain_velocity}

\item{ligand_target_matrix}{NicheNet ligand-target data file.}

\item{color_low}{Hex code of lower limit}

\item{color_high}{Hex code of upper limit.}

\item{colorscale}{Name of a color palette to be passed to the palette argument of scale_fill_distiller. Overrides color_low and color_high.}

\item{n_top_targets}{Number of top likelihood target genes to visualize. Default 10.}
}
\value{
A gglot2 plot object.
}
\description{
'Visualize the ligands and their feature importances towards the velocity clusters.
}
